refactor: extract useTopStampIds to share feature flag logic
