+++ src/components/Main/StampPicker/composables/useStampList.ts
@@ -1,10 +1,8 @@
 import type { Ref } from 'vue'
 import { computed } from 'vue'
 
-import { useFeatureFlagSettings } from '/@/store/app/featureFlagSettings'
 import { useStampCategory } from '/@/store/domain/stampCategory'
-import { useStampHistory } from '/@/store/domain/stampHistory'
-import { useStampRecommendations } from '/@/store/domain/stampRecommendations'
+import { useTopStampIds } from '/@/store/domain/stampRecommendations'
 import { useStampPalettesStore } from '/@/store/entities/stampPalettes'
 import { useStampsStore } from '/@/store/entities/stamps'
 import type { StampId } from '/@/types/entity-ids'
@@ -14,17 +12,13 @@ import type { StampSet } from './useStampSetSelector'
 
 const useStampList = (currentStampSet: Ref<StampSet>) => {
   const { traQStampCategory, unicodeStampCategories } = useStampCategory()
-  const { recentStampIds } = useStampHistory()
-  const { stampRecommendations } = useStampRecommendations()
+  const { topStampIds } = useTopStampIds()
   const { stampsMap } = useStampsStore()
   const { stampPalettesMap } = useStampPalettesStore()
-  const { featureFlags } = useFeatureFlagSettings()
 
   const stampIds = computed((): readonly StampId[] => {
     if (currentStampSet.value.type === 'history') {
-      return featureFlags.value.stamp_recommendation.enabled
-        ? stampRecommendations.value
-        : recentStampIds.value
+      return topStampIds.value
     }
     if (currentStampSet.value.type === 'palette') {
       const id = currentStampSet.value.id
